+++ test/fixtures/ava/type.js
@@ -26,3 +26,51 @@ suite('isArray', t => {
     isArray = arr instanceof Array;
   });
 });
+
+suite.failing('isArray - demostrate bug', t => {
+  t.set('maxTime', 0.01);
+  t.set('minSamples', 2);
+
+  var arr = [];
+  var isArray = null;
+
+  t.cycle(() => {
+    t.truthy(isArray);
+  });
+
+  t.bench('Array.isArray', () => {
+    isArray = Array.isArray(arr);
+  });
+
+  t.bench('Object.prototype.toString.call', () => {
+    isArray = (Object.prototype.toString.call(arr) === 'object Array');
+  });
+
+  t.bench('InstanceOf', () => {
+    isArray = arr instanceof Array;
+  });
+});
+
+suite.failing('isArray - demostrate error', t => {
+  t.set('maxTime', 0.01);
+  t.set('minSamples', 2);
+
+  var arr = [];
+  var isArray = null;
+
+  /* t.cycle(() => {
+    t.truthy(isArray);
+  }); */
+
+  t.bench('Array.isArray', () => {
+    isArray = Array.issArray(arr);
+  });
+
+  t.bench('Object.prototype.toString.call', () => {
+    isArray = (Object.prototype.toString.call(arr) === 'object Array');
+  });
+
+  t.bench('InstanceOf', () => {
+    isArray = arr instanceof Array;
+  });
+});